+++ exported/aip_oldone_marble_head/aip_oldone_marble_head.scml
@@ -19,6 +19,9 @@
     <folder id="1" name="attacker">
         <file id="0" name="attacker/spike.png" width="300" height="300" pivot_x="0" pivot_y="1"/>
     </folder>
+    <folder id="2" name="head">
+        <file id="0" name="head/head.png" width="400" height="350" pivot_x="0" pivot_y="1"/>
+    </folder>
     <entity id="0" name="chesspiece">
         <animation id="0" name="aipJump" length="600">
             <mainline>
@@ -103,6 +106,90 @@
                 </key>
                 <key id="2" time="500" spin="0">
                     <object folder="1" file="0" x="-1.414333" y="108.571429" pivot_x="0.485762" pivot_y="0.5" angle="0" scale_x="0.710633" scale_y="0.710633"/>
+                </key>
+            </timeline>
+        </animation>
+        <animation id="3" name="aipStruggle" length="1600">
+            <mainline>
+                <key id="0">
+                    <object_ref id="0" name="head" folder="2" file="0" abs_x="4.367681" abs_y="14.726705" abs_pivot_x="0.487531" abs_pivot_y="0.198006" abs_angle="360" abs_scale_x="1" abs_scale_y="1" abs_a="1" timeline="0" key="0" z_index="0"/>
+                </key>
+                <key id="1" time="392">
+                    <object_ref id="0" timeline="0" key="1" z_index="0"/>
+                </key>
+                <key id="2" time="1032">
+                    <object_ref id="0" timeline="0" key="2" z_index="0"/>
+                </key>
+                <key id="3" time="1415">
+                    <object_ref id="0" timeline="0" key="3" z_index="0"/>
+                </key>
+            </mainline>
+            <timeline id="0" name="head">
+                <key id="0">
+                    <object folder="2" file="0" x="4.367681" y="14.726705" pivot_x="0.487531" pivot_y="0.198006"/>
+                </key>
+                <key id="1" time="392" spin="-1">
+                    <object folder="2" file="0" x="4.367681" y="14.726705" pivot_x="0.487531" pivot_y="0.198006" angle="23.026595" scale_y="1"/>
+                </key>
+                <key id="2" time="1032">
+                    <object folder="2" file="0" x="4.367681" y="14.726705" pivot_x="0.487531" pivot_y="0.198006" angle="336.734246"/>
+                </key>
+                <key id="3" time="1415" spin="-1">
+                    <object folder="2" file="0" x="4.367681" y="14.726705" pivot_x="0.487531" pivot_y="0.198006" angle="9.250157"/>
+                </key>
+            </timeline>
+        </animation>
+        <animation id="4" name="aipCollect" length="2000">
+            <mainline>
+                <key id="0">
+                    <object_ref id="0" name="head" folder="2" file="0" abs_x="4.367681" abs_y="14.726705" abs_pivot_x="0.487531" abs_pivot_y="0.198006" abs_angle="360" abs_scale_x="1" abs_scale_y="1" abs_a="1" timeline="0" key="0" z_index="0"/>
+                </key>
+                <key id="1" time="512">
+                    <object_ref id="0" timeline="0" key="1" z_index="0"/>
+                </key>
+                <key id="2" time="740">
+                    <object_ref id="0" timeline="0" key="2" z_index="0"/>
+                </key>
+                <key id="3" time="958">
+                    <object_ref id="0" timeline="0" key="3" z_index="0"/>
+                </key>
+                <key id="4" time="1155">
+                    <object_ref id="0" timeline="0" key="4" z_index="0"/>
+                </key>
+                <key id="5" time="1376">
+                    <object_ref id="0" timeline="0" key="5" z_index="0"/>
+                </key>
+                <key id="6" time="1585">
+                    <object_ref id="0" timeline="0" key="6" z_index="0"/>
+                </key>
+                <key id="7" time="1793">
+                    <object_ref id="0" timeline="0" key="7" z_index="0"/>
+                </key>
+            </mainline>
+            <timeline id="0" name="head">
+                <key id="0" spin="0">
+                    <object folder="2" file="0" x="4.367681" y="14.726705" pivot_x="0.487531" pivot_y="0.198006"/>
+                </key>
+                <key id="1" time="512" spin="0">
+                    <object folder="2" file="0" x="4.367681" y="14.726705" pivot_x="0.487531" pivot_y="0.198006" scale_x="1.245621" scale_y="0.878171"/>
+                </key>
+                <key id="2" time="740" spin="0">
+                    <object folder="2" file="0" x="4.367681" y="14.726705" pivot_x="0.487531" pivot_y="0.198006" scale_x="1.138588" scale_y="0.91077"/>
+                </key>
+                <key id="3" time="958" spin="0">
+                    <object folder="2" file="0" x="4.367681" y="14.726705" pivot_x="0.487531" pivot_y="0.198006" scale_x="1.245621" scale_y="0.878171"/>
+                </key>
+                <key id="4" time="1155" spin="0">
+                    <object folder="2" file="0" x="4.367681" y="14.726705" pivot_x="0.487531" pivot_y="0.198006" scale_x="1.150754" scale_y="0.949295"/>
+                </key>
+                <key id="5" time="1376" spin="0">
+                    <object folder="2" file="0" x="4.367681" y="14.726705" pivot_x="0.487531" pivot_y="0.198006" scale_x="1.245621" scale_y="0.878171"/>
+                </key>
+                <key id="6" time="1585" spin="0">
+                    <object folder="2" file="0" x="4.367681" y="14.726705" pivot_x="0.487531" pivot_y="0.198006" scale_x="1.158865" scale_y="0.97893"/>
+                </key>
+                <key id="7" time="1793" spin="0">
+                    <object folder="2" file="0" x="4.367681" y="14.726705" pivot_x="0.487531" pivot_y="0.198006" scale_x="1.245621" scale_y="0.878171"/>
                 </key>
             </timeline>
         </animation>
